+++ app/src/main/java/com/hhp227/yu_minigroup/activity/CreateArticleActivity.java
@@ -16,6 +16,7 @@ import android.widget.Toast;
 
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.FileProvider;
 import androidx.exifinterface.media.ExifInterface;
@@ -25,13 +26,17 @@ import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.google.android.material.snackbar.Snackbar;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 import com.hhp227.yu_minigroup.R;
 import com.hhp227.yu_minigroup.adapter.WriteListAdapter;
 import com.hhp227.yu_minigroup.app.AppController;
 import com.hhp227.yu_minigroup.app.EndPoint;
 import com.hhp227.yu_minigroup.databinding.ActivityCreateArticleBinding;
+import com.hhp227.yu_minigroup.dto.ArticleItem;
 import com.hhp227.yu_minigroup.dto.YouTubeItem;
 import com.hhp227.yu_minigroup.helper.BitmapUtil;
 import com.hhp227.yu_minigroup.helper.PreferenceManager;
@@ -49,9 +54,9 @@ import java.util.*;
 public class CreateArticleActivity extends AppCompatActivity {
     private static final String TAG = CreateArticleActivity.class.getSimpleName();
 
-    private boolean mIsAdmin;
+    private String mGrpId, mGrpNm, mGrpImg, mCurrentPhotoPath, mCookie;
 
-    private String mGrpId, mGrpNm, mGrpImg, mCurrentPhotoPath, mCookie, mKey;
+    private String mArtlNum, mTitle, mContent, mGrpKey, mArtlKey;
 
     private List<String> mImageList;
 
@@ -82,11 +87,20 @@ public class CreateArticleActivity extends AppCompatActivity {
         mPreferenceManager = AppController.getInstance().getPreferenceManager();
         mCookie = AppController.getInstance().getCookieManager().getCookie(EndPoint.LOGIN_LMS);
         mProgressDialog = new ProgressDialog(this);
-        mIsAdmin = getIntent().getBooleanExtra("admin", false);
         mGrpId = getIntent().getStringExtra("grp_id");
         mGrpNm = getIntent().getStringExtra("grp_nm");
         mGrpImg = getIntent().getStringExtra("grp_img");
-        mKey = getIntent().getStringExtra("key");
+
+        //
+        mArtlNum = getIntent().getStringExtra("artl_num");
+        mTitle = getIntent().getStringExtra("sbjt");
+        mContent = getIntent().getStringExtra("txt");
+        mImageList = getIntent().getStringArrayListExtra("img");
+        mYouTubeItem = getIntent().getParcelableExtra("vid");
+        mGrpKey = getIntent().getStringExtra("grp_key");
+        mArtlKey = getIntent().getStringExtra("artl_key");
+
+        //
         mCameraPickActivityResultLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
             if (result.getResultCode() == RESULT_OK && result.getData() != null) {
                 Uri fileUri = result.getData().getData();
@@ -126,18 +140,27 @@ public class CreateArticleActivity extends AppCompatActivity {
                 mAdapter.notifyItemInserted(mContents.size() - 1);
             }
         });
+        Map<String, Object> headerMap = new HashMap<>();
 
         setContentView(mBinding.getRoot());
         setSupportActionBar(mBinding.toolbar);
         if (getSupportActionBar() != null) {
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         }
-        mAdapter.addHeaderView(new HashMap<>());
+        headerMap.put("title", mTitle);
+        headerMap.put("content", mContent);
+        mAdapter.addHeaderView(headerMap);
         mBinding.rvWrite.setLayoutManager(new LinearLayoutManager(this));
         mBinding.rvWrite.setAdapter(mAdapter);
         mBinding.llImage.setOnClickListener(this::showContextMenu);
         mBinding.llVideo.setOnClickListener(this::showContextMenu);
         mProgressDialog.setCancelable(false);
+        if (mImageList != null && mImageList.size() > 0) {
+            mContents.addAll(mImageList);
+            mAdapter.notifyDataSetChanged();
+        }
+        if (mYouTubeItem != null)
+            mContents.add(mYouTubeItem.position + 1, mYouTubeItem);
     }
 
     @Override
@@ -175,7 +198,11 @@ public class CreateArticleActivity extends AppCompatActivity {
                     if (mContents.size() > 1) {
                         int position = 1;
 
-                        if (mContents.get(position) instanceof Bitmap) {////////////// 리팩토링 요망
+                        if (mContents.get(position) instanceof String) {
+                            String image = (String) mContents.get(position);
+
+                            uploadProcess(position, image, false);
+                        } else if (mContents.get(position) instanceof Bitmap) {////////////// 리팩토링 요망
                             Bitmap bitmap = (Bitmap) mContents.get(position);// 수정
 
                             uploadImage(position, bitmap); // 수정
@@ -185,7 +212,7 @@ public class CreateArticleActivity extends AppCompatActivity {
                             uploadProcess(position, youTubeItem.videoId, true);
                         }
                     } else
-                        actionSend(mGrpId, title, content);
+                        actionSend(title, content);
                 } else
                     Snackbar.make(getCurrentFocus(), (title.isEmpty() ? "제목" : "내용") + "을 입력하세요.", Snackbar.LENGTH_LONG).setAction("Action", null).show();
                 return true;
@@ -215,8 +242,9 @@ public class CreateArticleActivity extends AppCompatActivity {
 
         switch (item.getGroupId()) {
             case 0:
-                if (mContents.get(item.getItemId()) instanceof YouTubeItem)
+                if (mContents.get(item.getItemId()) instanceof YouTubeItem) {
                     mYouTubeItem = null;
+                }
                 mContents.remove(item.getItemId());
                 mAdapter.notifyItemRemoved(item.getItemId());
                 return true;
@@ -252,7 +280,6 @@ public class CreateArticleActivity extends AppCompatActivity {
                 else {
                     Intent ysIntent = new Intent(getApplicationContext(), YouTubeSearchActivity.class);
 
-                    ysIntent.putExtra("type", 0);
                     mYouTubeSearchActivityResultLauncher.launch(ysIntent);
                 }
                 return true;
@@ -296,13 +323,13 @@ public class CreateArticleActivity extends AppCompatActivity {
         Volley.newRequestQueue(this).add(multipartRequest);
     }
 
-    private void uploadProcess(int position, String imageSrc, boolean isYoutube) { // 추가
+    private void uploadProcess(int position, String imageUrl, boolean isYoutube) { // 추가
         if (!isYoutube)
-            mImageList.add(imageSrc);
+            mImageList.add(imageUrl);
         mProgressDialog.setProgress((int) ((double) (position) / (mContents.size() - 1) * 100));
         try {
-            String test = (isYoutube ? "<p><embed title=\"YouTube video player\" class=\"youtube-player\" autostart=\"true\" src=\"//www.youtube.com/embed/" + imageSrc + "?autoplay=1\"  width=\"488\" height=\"274\"></embed><p>" // 유튜브 태그
-                    : ("<p><img src=\"" + imageSrc + "\" width=\"488\"><p>")) + (position < mContents.size() - 1 ? "<br>": "");
+            String test = (isYoutube ? "<p><embed title=\"YouTube video player\" class=\"youtube-player\" autostart=\"true\" src=\"//www.youtube.com/embed/" + imageUrl + "?autoplay=1\"  width=\"488\" height=\"274\"></embed><p>" // 유튜브 태그
+                    : ("<p><img src=\"" + imageUrl + "\" width=\"488\"><p>")) + (position < mContents.size() - 1 ? "<br>": "");
 
             mMakeHtmlContents.append(test);
             if (position < mContents.size() - 1) {
@@ -314,6 +341,10 @@ public class CreateArticleActivity extends AppCompatActivity {
                     Bitmap bitmap = (Bitmap) mContents.get(position);
 
                     uploadImage(position, bitmap);
+                } else if (mContents.get(position) instanceof String) {
+                    String imageSrc = (String) mContents.get(position);
+
+                    uploadProcess(position, imageSrc, false);
                 } else if (mContents.get(position) instanceof YouTubeItem) {
                     YouTubeItem youTubeItem = (YouTubeItem) mContents.get(position);
 
@@ -323,7 +354,7 @@ public class CreateArticleActivity extends AppCompatActivity {
                 String title = (String) mAdapter.getTextMap().get("title");
                 String content = (!TextUtils.isEmpty(mAdapter.getTextMap().get("content").toString()) ? Html.toHtml((Spanned) mAdapter.getTextMap().get("content"), Html.TO_HTML_PARAGRAPH_LINES_INDIVIDUAL) + "<p><br data-mce-bogus=\"1\"></p>" : "") + mMakeHtmlContents.toString();
 
-                actionSend(mGrpId, title, content);
+                actionSend(title, content);
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -332,7 +363,15 @@ public class CreateArticleActivity extends AppCompatActivity {
         }
     }
 
-    private void actionSend(String grpId, String title, String content) {
+    private void actionSend(String title, String content) {
+        if (getIntent().getIntExtra("type", -1) == 0) {
+            actionCreate(title, content);
+        } else {
+            actionUpdate(title, content);
+        }
+    }
+
+    private void actionCreate(final String title, final String content) {
         String tagStringReq = "req_send";
         StringRequest stringRequest = new StringRequest(Request.Method.POST, EndPoint.WRITE_ARTICLE, response -> {
             hideProgressDialog();
@@ -341,17 +380,6 @@ public class CreateArticleActivity extends AppCompatActivity {
                 boolean error = jsonObject.getBoolean("isError");
 
                 if (!error) {
-                    /*Intent intent = new Intent(CreateArticleActivity.this, GroupActivity.class);
-
-                    intent.putExtra("admin", mIsAdmin);
-                    intent.putExtra("grp_id", grpId);
-                    intent.putExtra("grp_nm", mGrpNm);
-                    intent.putExtra("grp_img", mGrpImg);
-                    intent.putExtra("key", mKey);
-
-                    // 이전 Activity 초기화
-                    intent.setFlags(Intent.FLAG_ACTIVITY_TASK_ON_HOME | Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                    startActivity(intent);*/
                     setResult(RESULT_OK);
                     finish();
                     Toast.makeText(getApplicationContext(), "전송완료", Toast.LENGTH_LONG).show();
@@ -379,7 +407,49 @@ public class CreateArticleActivity extends AppCompatActivity {
                 Map<String, String> params = new HashMap<>();
 
                 params.put("SBJT", title);
-                params.put("CLUB_GRP_ID", grpId);
+                params.put("CLUB_GRP_ID", mGrpId);
+                params.put("TXT", content);
+                return params;
+            }
+        };
+        AppController.getInstance().addToRequestQueue(stringRequest, tagStringReq);
+    }
+
+    private void actionUpdate(final String title, final String content) {
+        String tagStringReq = "req_send";
+        StringRequest stringRequest = new StringRequest(Request.Method.POST, EndPoint.MODIFY_ARTICLE, response -> {
+            hideProgressDialog();
+            try {
+                Intent intent = new Intent(CreateArticleActivity.this, ArticleActivity.class);
+
+                Toast.makeText(getApplicationContext(), "수정완료", Toast.LENGTH_LONG).show();
+                setResult(RESULT_OK, intent);
+                finish();
+            } catch (Exception e) {
+                Log.e(TAG, e.getMessage());
+            } finally {
+                initFirebaseData();
+            }
+        }, error -> {
+            VolleyLog.e(error.getMessage());
+            Toast.makeText(getApplicationContext(), error.getMessage(), Toast.LENGTH_LONG).show();
+            hideProgressDialog();
+        }) {
+            @Override
+            public Map<String, String> getHeaders() {
+                Map<String, String> headers = new HashMap<>();
+
+                headers.put("Cookie", mCookie);
+                return headers;
+            }
+
+            @Override
+            protected Map<String, String> getParams() {
+                Map<String, String> params = new HashMap<>();
+
+                params.put("CLUB_GRP_ID", mGrpId);
+                params.put("ARTL_NUM", mArtlNum);
+                params.put("SBJT", title);
                 params.put("TXT", content);
                 return params;
             }
@@ -437,7 +507,35 @@ public class CreateArticleActivity extends AppCompatActivity {
         map.put("content", TextUtils.isEmpty(mAdapter.getTextMap().get("content").toString()) ? null : mAdapter.getTextMap().get("content").toString());
         map.put("images", mImageList);
         map.put("youtube", mYouTubeItem);
-        databaseReference.child(mKey).push().setValue(map);
+        databaseReference.child(mGrpKey).push().setValue(map);
+    }
+
+    private void initFirebaseData() {
+        DatabaseReference databaseReference = FirebaseDatabase.getInstance().getReference("Articles");
+
+        updateArticleDataToFirebase(databaseReference.child(mGrpKey).child(mArtlKey));
+    }
+
+    private void updateArticleDataToFirebase(DatabaseReference query) {
+        query.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                ArticleItem articleItem = dataSnapshot.getValue(ArticleItem.class);
+
+                if (articleItem != null) {
+                    articleItem.setTitle((String) mAdapter.getTextMap().get("title"));
+                    articleItem.setContent(TextUtils.isEmpty(mAdapter.getTextMap().get("content").toString()) ? null : mAdapter.getTextMap().get("content").toString());
+                    articleItem.setImages(mImageList.isEmpty() ? null : mImageList);
+                    articleItem.setYoutube(mYouTubeItem);
+                    query.getRef().setValue(articleItem);
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+                Log.e("파이어베이스", databaseError.getMessage());
+            }
+        });
     }
 
     private void showProgressDialog() {
